Prodigy/task2/requirements.txt
torch>=1.9.0
transformers>=4.21.0
diffusers>=0.18.0
accelerate>=0.12.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.0
huggingface-hub>=0.10.0
ftfy>=6.0.0
regex>=2022.1.18

